# Backend/src/main/java/web/labs/work/dto/TaskDto.java
package web.labs.work.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDate;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class TaskDto {
    private Long id;
    private String name;
    private String description;
    private LocalDate endDate;
    private boolean completed;
    private Long projectId;
}