# Backend/src/main/java/web/labs/work/dto/TeamDto.java
package web.labs.work.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class TeamDto {
    Integer id;
    String name;
    String description;
    Long projectId;
}